requirements.txt
fastapi
bs4
langchain
langchain_groq
langchain_community
python-dotenv
requests